% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_transformations.R
\name{alr}
\alias{alr}
\title{alr}
\usage{
alr(x)
}
\arguments{
\item{x}{vector of compositions doesn't actually have to sum = 1}
}
\value{
additive log ratio transformed variable
}
\description{
transformation
sum(x) = 1 & length(x) = n -> y length(y) = n - 1
The last value is made to be the 'reference' element
}
