% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_transformations.R
\name{alrinv}
\alias{alrinv}
\title{alrinv}
\usage{
alrinv(y)
}
\arguments{
\item{y}{vector of compositions that have been alr}
}
\value{
additive log ratio transformed variable
}
\description{
ALN transformation inverse
sum(x) = 1 & length(x) = n -> y length(y) = n - 1
}
