% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grooming.R
\name{apply_grooming_rule}
\alias{apply_grooming_rule}
\title{apply_grooming_rule}
\usage{
apply_grooming_rule(df, index)
}
\arguments{
\item{df}{the df we are manipulating}

\item{index}{a conditional index length(index) == nrow(df). TRUE means it stays, FALSE means it is being excluded from}
}
\value{
a modified version of df with the FALSE values from 'index' removed
}
\description{
apply_grooming_rule
}
\details{
will remove rows from a data frame based on the conditional vector index.
}
