% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_params.R
\name{basic_weight}
\alias{basic_weight}
\title{basic_weight calculate weight for a given length based on allometric length weight relationship with bias correction
bias correction follows CASAL which is applied in age-structured models.}
\usage{
basic_weight(a, b, l, cv, distribution)
}
\arguments{
\item{a}{a parameter in growht function}

\item{b}{exponent}

\item{l}{length}

\item{cv}{cv from the age-length distribution}

\item{distribution}{of age-length distribution}
}
\value{
weight, units will depend on a and b
}
\description{
basic_weight calculate weight for a given length based on allometric length weight relationship with bias correction
bias correction follows CASAL which is applied in age-structured models.
}
