% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/recruit_funs.R
\name{BH_non_equil}
\alias{BH_non_equil}
\title{BH_non_equil}
\usage{
BH_non_equil(SSB, alpha, beta)
}
\arguments{
\item{SSB}{Spawning stock biomass}

\item{alpha}{classic params alpha BH param}

\item{beta}{classic params beta BH param}
}
\value{
Number of recruits acording to the Beverton holt relationship
}
\description{
Beverton holt stock recuit relationship parameterised with a and b instead of steepness
}
