% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_helper_funs.R
\name{check_tmb_convergence}
\alias{check_tmb_convergence}
\title{check_tmb_convergence}
\usage{
check_tmb_convergence(obj, delta = 0.001)
}
\arguments{
\item{obj}{An optimised list that has been build by MakeAdFun}

\item{delta}{Gradient threshold for defining a converged model}
}
\description{
use TMB object to check gradients and if the hessian is definite positive
}
\author{
C.Marsh
}
