% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_AF.R
\name{compare_AF}
\alias{compare_AF}
\title{compare_AF between an abm output and CASAL model run}
\usage{
compare_AF(
  cas_AFs,
  abm_output = abm,
  abm_report_lab = "age_freq_step1",
  year_of_interest,
  region,
  plotit = F,
  ylim = NULL
)
}
\arguments{
\item{cas_AFs}{a list object that has been generated from the function extract.AFs() or extract.step.AFs()}

\item{abm_output}{an ABM model output}

\item{year_of_interest}{}

\item{region}{assumes regions are consistently labelled between Casal model and ABM}

\item{plotit}{whether you want this function to create a ggplot or return a long form data frame}

\item{abm}{report label abm_report_lab}
}
\description{
compare_AF between an abm output and CASAL model run
}
