% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_params.R
\name{cv_by_length}
\alias{cv_by_length}
\title{cv_by_length  does a length based interpolation of cv for the growth function. This is also done in CASAL when by_length = T.}
\usage{
cv_by_length(cv1, cv2, mean_length_at_age)
}
\arguments{
\item{cv1}{cv for first mean length}

\item{cv2}{cv for last mean length}

\item{mean_length_at_age}{mean length at age from a growth function or empirical}
}
\value{
vector of cvs for each mean length
}
\description{
cv_by_length  does a length based interpolation of cv for the growth function. This is also done in CASAL when by_length = T.
}
