% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_funs.R
\name{decdeg2deg}
\alias{decdeg2deg}
\title{decdeg2deg}
\usage{
decdeg2deg(x)
}
\arguments{
\item{x}{<scalar> format decimal degrees}
}
\value{
degrees:minutes:seconds
}
\description{
convert from  decimal degrees -> degrees degrees:minutes:seconds
}
\examples{
\dontrun{
deg2decdeg(paste(-43,50,23,sep = ":"))
}
}
