% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_funs.R
\name{deg2decdeg}
\alias{deg2decdeg}
\title{deg2decdeg}
\usage{
deg2decdeg(x)
}
\arguments{
\item{x}{<scalar> format degrees:minutes:seconds}
}
\value{
dec_degrees
}
\description{
convert from degrees degrees:minutes:seconds -> decimal degrees
}
\examples{
\dontrun{
deg2decdeg(paste(-43,50,23,sep = ":"))
}
}
