% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_funs.R
\name{DistanceLongLat}
\alias{DistanceLongLat}
\title{DistanceLongLat}
\usage{
DistanceLongLat(long1, long2, lat1, lat2, metres = F)
}
\arguments{
\item{long1}{longitude first point decimal degrees > 180 are WESTINGS}

\item{long2}{longitude second point decimal degrees > 180 are WESTINGS}

\item{lat1}{lattitude first point decimal degrees < 0 Southings}

\item{lat2}{lattitude second point decimal degrees < 0 Southings}

\item{metres}{bool, F = nautical miles, T = metres}
}
\value{
a distance
}
\description{
calculate distance between a set of coordinates (as the crow flies)
}
