% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_transformations.R
\name{draw_from_radius}
\alias{draw_from_radius}
\title{draw_from_radius
description Stan starts initial values by drawing random values within a radius (+/-2) around zero
this script will have a fun funs, that will draw a range of paraemters from this, that you
can feed into your model to run deterministically to identify problem paraemters when you}
\usage{
draw_from_radius(
  n = 10,
  lower_bound = -Inf,
  upper_bound = Inf,
  minus_init_radius = -2,
  plus_init_radius = 2
)
}
\arguments{
\item{n}{number of starting values}

\item{lower_bound}{Lower bound of parameter}

\item{upper_bound}{Upper bound of parameter}

\item{minus_init_radius}{lower bound of radius to draw unconstrained starting values}

\item{plus_init_radius}{upper bound of radius to draw unconstrained starting values}
}
\value{
start_values for model
}
\description{
draw_from_radius
description Stan starts initial values by drawing random values within a radius (+/-2) around zero
this script will have a fun funs, that will draw a range of paraemters from this, that you
can feed into your model to run deterministically to identify problem paraemters when you
}
