% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_helper_funs.R
\name{eigen_decomp_covariance}
\alias{eigen_decomp_covariance}
\title{eigen_decomp_covariance}
\usage{
eigen_decomp_covariance(
  covariance_matrix,
  param_labels = NULL,
  delta = .Machine$double.eps
)
}
\arguments{
\item{covariance_matrix}{symetric covariance matrix}

\item{param_labels}{vector of param labels (optional)}

\item{delta}{a cut off value for 'poorly' defined parameters.}
}
\value{
data frame of eiegen values for the matrix and index of good and bad parameters based on delta
}
\description{
Do an eigen decomposition to look at poorly estimated parameters from MLE fit
}
