% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.csl.file.R
\name{extract.csl.file}
\alias{extract.csl.file}
\title{extract.csl.file an alternative to casal::extract.csl.file() which I had issues with.}
\usage{
extract.csl.file(file, path = "")
}
\arguments{
\item{file}{casal output filename}

\item{path}{path to the filename}
}
\value{
list of casal config files.
}
\description{
extract.csl.file an alternative to casal::extract.csl.file() which I had issues with.
}
