% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.multiple.free.params.R
\name{extract.multiple.free.params}
\alias{extract.multiple.free.params}
\title{extract.multiple.free.params}
\usage{
extract.multiple.free.params(file, path = "")
}
\arguments{
\item{file}{character of filename}

\item{path}{where the file can be found}
}
\value{
a list of lists, each element of the top level list is for each -i run
}
\description{
extract.multiple.free.params
}
\details{
extracts a list of model free parameters that have output from a CASAL model using -i format, or a profile -p. This function assumes you are reporting free parameters (I think this is the default)
}
\examples{
\dontrun{
Add example usage
}
}
