% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.multiple.quantities.R
\name{extract.multiple.quantities}
\alias{extract.multiple.quantities}
\title{extract.multiple.quantities}
\usage{
extract.multiple.quantities(file, path = "")
}
\arguments{
\item{file}{character of filename}

\item{path}{where the file can be found}
}
\value{
a list of lists, each element of the top level list is for each -i run
}
\description{
extract.multiple.quantities
}
\details{
extracts a list of model quantities that have output from a CASAL model using -i format
}
\examples{
\dontrun{
file = "test_est.out"
csl_dir = "../csl"
}
}
