% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.state.R
\name{extract.state}
\alias{extract.state}
\title{extract.state extract State at the end of year}
\usage{
extract.state(file, path = "")
}
\arguments{
\item{file}{casal output filename}

\item{path}{path to the filename}
}
\value{
list of AFs
}
\description{
extract.state extract State at the end of year
}
