% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract.step.AFs.R
\name{extract.step.AFs}
\alias{extract.step.AFs}
\title{extract.step.AFs State at the end of year and a step}
\usage{
extract.step.AFs(file, step = 1, path = "")
}
\arguments{
\item{file}{casal output filename}

\item{path}{path to the filename}
}
\value{
list of AFs
}
\description{
extract.step.AFs State at the end of year and a step
}
