% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmb_helper_funs.R
\name{fix_pars}
\alias{fix_pars}
\title{fix_pars}
\usage{
fix_pars(
  par_list,
  pars_to_exclude,
  vec_pars_to_adjust = NULL,
  vec_elements_to_exclude = NULL
)
}
\arguments{
\item{par_list}{a named list that you give to the par argument in the MakeADFun}

\item{pars_to_exclude}{a vector of strings with names of parmeters you want to FIX in the objective object.}

\item{vec_pars_to_adjust}{a vector string of parameter labels that we want to exclude certain elements.}

\item{vec_elements_to_exclude}{a list with number of elements = length(vec_pars_to_adjust). each list element 
contains a vector of elements that we want to exclude from estimation.}
}
\value{
a list of factors used in the MakeADFun function
}
\description{
TMB helper function this function returns a list of factors used in the map argument of the MakeADFun function
}
\author{
C.Marsh
}
