% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surplus_production_funs.R
\name{general_surplus_production_model_sigma}
\alias{general_surplus_production_model_sigma}
\title{general_surplus_production_model_sigma}
\usage{
general_surplus_production_model_sigma(
  seed = 123,
  n_t,
  B1,
  m,
  r,
  k,
  catches = NULL,
  exploitation = NULL,
  obs_std,
  catchability,
  observation_likelihood = 1,
  process_error = NULL,
  process_std = NULL,
  n_obs_series = 1
)
}
\arguments{
\item{seed}{seed for random number generator}

\item{n_t}{<integer> number of time-steps to run the model}

\item{B1}{<real> starting exploited biomass}

\item{m}{<real> shape parameter for the surplus production curve (m = 2 = schafer)}

\item{r}{<real> intrinsic growth parameter.}

\item{k}{<real> carrying capacity parameter.}

\item{catches}{<vector> observed catches, defines number of years, so if you have a period of no catches set values of 0, number of years = length(catches+ 1)}

\item{exploitation}{<vector> observed catches, defines number of years, so if you have a period of no catches set values of 0, number of years = length(catches+ 1}

\item{obs_std}{<real> Standard deviation which is used to generate simulated observations.}

\item{catchability}{<vector> a scalar to generate the relative index observation}

\item{observation_likelihood}{<int> likelihood type, 1 = lognormal, 2 = normal, 3 = ...}

\item{process_error}{(optional) time specific time deviations (epsilon_t,p) for each time step}

\item{process_std}{(optional) standard deviation for process deviations corrections}

\item{n_obs_series}{number relative indicies}
}
\value{
Returns a list of model quantities and observations
}
\description{
that can be used as an operating model
Different from the above uses a standard deviation for process error instead of CV and also allow for fishing to be input as both catch and exploitation.
}
