% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_math_funs.R
\name{geometric.mean}
\alias{geometric.mean}
\title{Utility functions for generic maths, mostly the same function but ignoring NA's}
\usage{
geometric.mean(x)
}
\arguments{
\item{x}{vecor of strictly positive values}
}
\value{
geometric mean with na's ignored
}
\description{
calculate geometric mean when you have na's in vector
}
