% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_df_quantiles.R
\name{get_df_quantiles}
\alias{get_df_quantiles}
\title{get_df_quantiles
get quanitles from a data.frame for 'y_value', grouped by group_vars using purrr and dplyr}
\usage{
get_df_quantiles(df, group_vars, y_value, quants = c(0.025, 0.5, 0.975))
}
\arguments{
\item{df}{a dataframe with columns in group_vars and y_value}

\item{group_vars}{a vector of strings specifying the grouping variables}

\item{y_value}{a string specifying the column to calculate the quanitles for}

\item{quants}{numeric vector of values between 0-1 which define the quantiles to calculate.}
}
\value{
a dataframe of quantiles
}
\description{
get_df_quantiles
get quanitles from a data.frame for 'y_value', grouped by group_vars using purrr and dplyr
}
