% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_invertable.R
\name{is_invertable}
\alias{is_invertable}
\title{is_invertable test whether a matrix is invertable, used to test for hessian or covariance is available}
\usage{
is_invertable(m)
}
\arguments{
\item{m}{matrix to be tested}
}
\value{
boolean true if invertable and false If not
}
\description{
is_invertable test whether a matrix is invertable, used to test for hessian or covariance is available
}
\author{
Craig Marsh
}
