% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_transformations.R
\name{lb_constrain}
\alias{lb_constrain}
\title{lb_constrain convert x(-Inf, Inf) -> y(Lb, Inf]}
\usage{
lb_constrain(x, Lb)
}
\arguments{
\item{x}{value to transform}

\item{Lb}{Lower bound of value}
}
\value{
constrained (natural space) value
}
\description{
lb_constrain convert x(-Inf, Inf) -> y(Lb, Inf]
}
