% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_funs.R
\name{lnorm_prior}
\alias{lnorm_prior}
\title{lnorm_prior}
\usage{
lnorm_prior(X, expectation, sigma, negative = TRUE)
}
\arguments{
\item{X}{<double>: value/'s to calculate the prior for}

\item{expectation}{<double>: The expectation of the distribution, this is not the mu parameter because for the lognormal distribution the expectation is not the expectation parameter.}

\item{sigma}{<double>: The standard deviation of the distribution (in log space)}

\item{negative}{<bool>: Return the negative log-likelihod if True else if FALSE return just the log-likelihood}
}
\description{
calculate the log-likelihood contribution for the log-normal prior in CASAL
}
