% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectivity_ogives.R
\name{logis_sel}
\alias{logis_sel}
\title{logis_sel Logistic selectivity}
\usage{
logis_sel(bins, a50, ato95)
}
\arguments{
\item{bins}{could be age or length}

\item{a50}{bin which equates to selectivy being at 0.5}

\item{ato95}{bins between the selectivty 0.5 and 0.95}
}
\value{
selectivity for each age.
}
\description{
logis_sel Logistic selectivity
}
