% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_transformations.R
\name{logit}
\alias{logit}
\title{logit logistic transformation same as qlogis}
\usage{
logit(u)
}
\arguments{
\item{u}{[0,1]}
}
\value{
transformed value
}
\description{
logit logistic transformation same as qlogis
}
\author{
C.Marsh
}
