% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lognormal_funs.R
\name{lognormal_CI}
\alias{lognormal_CI}
\title{lognormal_CI Calculate upper and lower bounds for the lognormal distribution based with expectation and cv}
\usage{
lognormal_CI(expectation, cv, CI = 0.95)
}
\arguments{
\item{expectation}{The expectation of the distribution, this is not the mu parameter because for the lognormal distribution the expectation is not the mu parameter.}

\item{cv}{The standard deviation of the lognormal distribution}

\item{CI}{level of confidence (units are proportions not percentage i.e. 0.95 for 95 CI)}
}
\value{
a list with upper and lower elements
}
\description{
lognormal_CI Calculate upper and lower bounds for the lognormal distribution based with expectation and cv
}
