% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_funs.R
\name{long2UTM}
\alias{long2UTM}
\title{long2UTM}
\usage{
long2UTM(long)
}
\arguments{
\item{long}{<scalar> longitude, to get UTM zone}
}
\value{
UTM zone
}
\description{
returns a zone for UTM based on longitude. Expects long can be either [-180, 180] or [0,360]
}
