% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial_funs.R
\name{LongLatToUTM}
\alias{LongLatToUTM}
\title{LongLatToUTM}
\usage{
LongLatToUTM(df, zone = NULL)
}
\arguments{
\item{df}{<data.frame> that has at least two columns labelled long lat}

\item{zone}{<integer> zone for UTM.}
}
\value{
SpatialDataFrame
}
\description{
takes a data frame with lat and long in decimal degree format. Then
returns a spatial dataframe that is in UTM format.
}
