% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_transformations.R
\name{lub_constrain}
\alias{lub_constrain}
\title{lub_constrain convert x(-Inf, Inf) -> y(Lb, Ub]}
\usage{
lub_constrain(x, Lb, Ub)
}
\arguments{
\item{x}{value to transform}

\item{Lb}{Lower bound of value}

\item{Ub}{Upper bound of value}
}
\value{
unconstrained value
}
\description{
lub_constrain convert x(-Inf, Inf) -> y(Lb, Ub]
}
