% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_transformations.R
\name{lub_free}
\alias{lub_free}
\title{lub_free convert y(Lb, Ub] -> x(-Inf, Inf)}
\usage{
lub_free(y, Lb, Ub)
}
\arguments{
\item{y}{value to transform}

\item{Lb}{Lower bound of value}

\item{Ub}{Upper bound of value}
}
\value{
unconstrained value
}
\description{
lub_free convert y(Lb, Ub] -> x(-Inf, Inf)
}
