% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_math_funs.R
\name{Mean}
\alias{Mean}
\title{overloading function to calculate arithmic mean when you have na's in vector}
\usage{
Mean(..., na.rm = T)
}
\arguments{
\item{...}{same parameters as mean(...)}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
mean with na's ignored
}
\description{
overloading function to calculate arithmic mean when you have na's in vector
}
