% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_math_funs.R
\name{Median}
\alias{Median}
\title{overloading function to calculate median when you have na's in vector}
\usage{
Median(..., na.rm = T)
}
\arguments{
\item{...}{same parameters as median(...)}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
median with na's ignored
}
\description{
overloading function to calculate median when you have na's in vector
}
