% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_math_funs.R
\name{normalise}
\alias{normalise}
\title{scale each element by the arithmic mean}
\usage{
normalise(vector, scalar = 1, na.rm = T)
}
\arguments{
\item{vector}{numeric vector to normalise}

\item{scalar}{multiplier to change average values}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
vector of normalised values to have mean = scalar
}
\description{
scale each element by the arithmic mean
}
\details{
na's will be ignored
}
