% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_CASAL_profile.R
\name{plot_CASAL_profile}
\alias{plot_CASAL_profile}
\title{plot_CASAL_profile}
\usage{
plot_CASAL_profile(
  file,
  path,
  profiled_param,
  objective_function_components = "all",
  element = NULL,
  ask_for_obj_labels = FALSE,
  likelihood_cut_off = 100,
  custom_xlim = NULL
)
}
\arguments{
\item{file}{file name from profiled outptu.}

\item{path}{path for profiled output}

\item{objective_function_components}{either specify 'all' which will plot all likelihood contributions, or a vector of characters for the contirbutions you want. If not sure how to specify these see ask_for_obj_labels}

\item{element}{an integer for when the profiled_param is a vector param.}

\item{ask_for_obj_labels}{boolean if True the function will print the values available for objective_function_components}

\item{likelihood_cut_off}{ylim for likelihood to remove large values obscuring the trend of the profile}
}
\value{
ggplot and data frame
}
\description{
This function
}
\examples{
\dontrun{

setwd("profileplots")
## ask for liklihood components
plot_CASAL_profile(file = '2021_profile_R0.casal', path = getwd(), profiled_param, objective_function_components = "all", element = NULL, ask_for_obj_labels = T, likelihood_cut_off = 100, custom_xlim = c(0,2e7))
## now plot it
my_plot = plot_CASAL_profile(file = '2021_profile_R0.casal', path = getwd(), profiled_param = "initialization.R0", objective_function_components = c("BP_SPUE", "REC_CPUE", "PS_AGE","REC_AGE", "ST_AGE", "prior_on_initialization.R0"), element = NULL, ask_for_obj_labels = FALSE, likelihood_cut_off = 100, custom_xlim = c(0,2e7))
my_plot = plot_CASAL_profile(file = '2021_profile_rec_ato95.casal', path = getwd(), profiled_param = "selectivity[Sel_REC].all", objective_function_components = c("BP_SPUE", "REC_CPUE", "PS_AGE","REC_AGE", "ST_AGE", "prior_on_initialization.R0"), element = 1, ask_for_obj_labels = FALSE, likelihood_cut_off = 100)
}

}
