% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp_resids.R
\name{plot_comp_resids}
\alias{plot_comp_resids}
\title{plot_comp_resids}
\usage{
plot_comp_resids(resids_matrix, show_legend = TRUE)
}
\arguments{
\item{resids_matrix}{a matrix[years,bins] that are standardised residuals from compositional data, rownames(resids_matrix) = years (y-axis), colnames(resids_matrix) = ages/length}
}
\value{
a ggplot bubble plot
}
\description{
A function that generates a plot of residuals for compositional data.
}
\examples{
\dontrun{
mpd = extract.mpd(path = system.file("extdata",package = "stockassessmenthelper"), file = "estimate.log")
colnames(mpd$fits$chatTANage$pearson.resids) = as.numeric(substring(colnames(mpd$fits$chatTANage$pearson.resids), first = 2))
plt = plot_comp_resids(resids_matrix = as.matrix(mpd$fits$chatTANage$pearson.resids), show_legend  = T)
}
}
\author{
C.Marsh
}
