% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_math_funs.R
\name{pow}
\alias{pow}
\title{pow exponential function, helpful when debugging C++ functions}
\usage{
pow(x, exponent)
}
\arguments{
\item{x}{base value}

\item{exponent}{the exponent to apply to base value}
}
\value{
x^exponent
}
\description{
pow exponential function, helpful when debugging C++ functions
}
