% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grooming.R
\name{record_grooming_rule}
\alias{record_grooming_rule}
\title{record_grooming_rule}
\usage{
record_grooming_rule(
  df,
  index,
  catch.col,
  record = NULL,
  rule,
  year.col = NULL,
  attribute = "catch"
)
}
\arguments{
\item{df}{the df we are manipulating}

\item{index}{a conditional index length(index) == nrow(df). TRUE means it stays, FALSE means it is being excluded from}

\item{catch.col}{column label for the catch variable}

\item{record}{a data.frame to append the summaries to.}

\item{rule}{the label for the grooming rule (string)}

\item{year.col}{column label for the year variable. Can be ommited, if supplied the Data.frame will break down catch and number of events removed from each grooming rule by year.}

\item{attribute}{if by year specify the attribute to record
\itemize{
  \item catch absoulte catch
  \item events number of events
  \item relative_catch percent change in catch
  \item relative_events percent change in events
}}
}
\value{
the record data frame with new entries for the grooming rule
}
\description{
record_grooming_rule
}
\details{
a utility function to record the effect of grooming rules on data sets. It will note remove the index from the dataset.
}
