% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_reweight.R
\name{run_reweight}
\alias{run_reweight}
\title{run_reweight}
\usage{
run_reweight(
  csl_path,
  est_csl_file,
  comp_obs,
  est_file,
  additional_casal_pars,
  tolerance = 0.01,
  iters = 5,
  min_years = 3
)
}
\arguments{
\item{csl_path}{directory path to where csl files are and where estimation will occur}

\item{est_csl_file}{label for the estimation csl file i.e. estimation.csl}

\item{comp_obs}{vector of strings that relate to block labels for the composition observations, that are consistent with csl file definitions i.e. EN_age}

\item{est_file}{estimated output from the casal estimation run}

\item{additional_casal_pars}{string of additional casal parameters i.e. "-q" to run quiet or "f my_stock_" if you have odd labeled csl files}

\item{tolerance}{difference from 1 to which we don't reweight the observation class}

\item{iters}{number of reweighting iterations to run, for you to manually tweak.}

\item{min_years}{<int> minimum years that are needed to weight this data.}
}
\value{
nothing, this function writes to files and estimates, you will want to look at the estimated output after to see the effects
}
\description{
A function for automatically runnning compositional reweighting code. It will overwrite the estimation.csl file so please read the details before running the function
}
\details{
before running this code you will need to save the original estimation.csl as this function will overwrite the est_csl_file file.
Applies the Method.TA1.8 method in francis
}
\examples{
\dontrun{
path = csl_path = "../csl"
file = est_csl_file = "K12014_estimation.csl"
est_file = "est_upd.out"
comp_obs = c("REC_AGE", "SN_AGE", "PS_AGE", "ST_AGE")
additional_casal_pars = "-f K12014_"
tolerance = 0.001
iters = 5
}
}
