% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_params.R
\name{schnute}
\alias{schnute}
\title{schnute}
\usage{
schnute(y1, y2, t1, t2, a, b, age)
}
\arguments{
\item{y1}{lenght at reference age t1}

\item{y2}{length at reference age t2}

\item{t1}{reference age for y1}

\item{t2}{reference age for y2}

\item{a}{schnute growth param}

\item{b}{schnute growth param}

\item{age}{age to calculate mean length for}
}
\value{
return teh mean length at age for the Schnute growth equation
}
\description{
general age-length growth formula
}
