% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_math_funs.R
\name{Sd}
\alias{Sd}
\title{overloading function to calculate standard deviation when you have na's in vector}
\usage{
Sd(..., na.rm = T)
}
\arguments{
\item{...}{same parameters as sd(...)}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
standard deviation with na's ignored
}
\description{
overloading function to calculate standard deviation when you have na's in vector
}
