% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_pars.R
\name{sim_pars}
\alias{sim_pars}
\title{sim_pars}
\usage{
sim_pars(
  dash_o_file_name,
  csl_path,
  plus_minus_for_bounds = 0.1,
  n_pars = 10,
  index_of_pars_fixed = NULL
)
}
\arguments{
\item{dash_o_file_name}{file name of a parameter file created from casal -r -o > par.file > run.out}

\item{csl_path}{the path to where csl files are and the dash_o_file_name}

\item{plus_minus_for_bounds}{this sets the parameter space used to simulate starting values from uniform distribution LB = param_val - param_val x plus_minus_for_bounds and UB = param_val + param_val x plus_minus_for_bounds}

\item{n_pars}{number of simulated parameters (rows)}

\item{index_of_pars_fixed}{if parameters are fixed by forcing lower_bound == upper_bound you will need to tell these functions this, i.e. index_of_pars_fixed = c(1,30) means the first and thirtith parameters area fixed and don't need to be simulated.}
}
\value{
will create a par file called sim_pars.out at csl_path will also return a matrix of the simulated value
}
\description{
simulate random start parameters using a uniform random variable for CASAL mdoels
}
\examples{
\dontrun{
csl_dir = "../csl"
sim_pars(dash_o_file_name = "par.out", csl_path = csl_dir, plus_minus_for_bounds= 0.5, n_pars = 10, index_of_pars_fixed = c(29:31))
# then run the following casal command
# casal -e -f K12014_ -i sim_pars.out > test_est.out   
multi_obj = extract.multiple.objective.functions(file = "test_est.out", path = csl_dir)
}
}
