% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surplus_production_funs.R
\name{simulate_SPM}
\alias{simulate_SPM}
\title{simulate_SPM}
\usage{
simulate_SPM(
  theta,
  r,
  m,
  K,
  exploitation,
  obs_cv,
  catchability,
  proc_std,
  proj_years,
  proj_u,
  process_error = NULL
)
}
\arguments{
\item{theta}{<real> starting exploited biomass}

\item{r}{<real> intrinsic growth parameter.}

\item{m}{<real> Shape parameter for the surplus production function}

\item{K}{<real> carrying capacity parameter.}

\item{exploitation}{<vector> Exploitation, defines number of years (length = n_t - 1), so if you have a period of no catches set values of 0}

\item{obs_cv}{<vector> CV's (for each year) for simulated observations}

\item{catchability}{<vector> a scalar to generate the relative index observation}

\item{proc_std}{(optional) time specific time deviations (epsilon_t,p) for each time step}

\item{proj_years}{int number of projection years}

\item{proj_u}{vector<proj_years> (length 1 or proj_years)}

\item{process_error}{<vector> CV's (for each year) for simulated observations}
}
\value{
Returns a list of model quantiteis and observations that can be used as an operating model
}
\description{
simulate a surplus production model, maybe fletchers?
}
