% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_params.R
\name{stretch}
\alias{stretch}
\title{stretch does a CV linear interpolation by age used in CASAL to define cv by age}
\usage{
stretch(cv1, cv2, min_age, max_age)
}
\arguments{
\item{cv1}{is the cv for min_age}

\item{cv2}{is the cv for max_age}

\item{min_age}{min age to calculate cvs for}

\item{max_age}{max age to calculate cvs for}
}
\value{
vector of cvs from min_age to max_age
}
\description{
stretch does a CV linear interpolation by age used in CASAL to define cv by age
}
