% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_math_funs.R
\name{Sum}
\alias{Sum}
\title{overloading function to calculate sum when you have na's in vector}
\usage{
Sum(..., na.rm = T)
}
\arguments{
\item{...}{same parameters as sum(...)}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
sum with na's ignored
}
\description{
overloading function to calculate sum when you have na's in vector
}
