% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surplus_production_funs.R
\name{surplus_production_millar}
\alias{surplus_production_millar}
\title{surplus_production_millar}
\usage{
surplus_production_millar(K, r, q, P1, obs_std, pro_std, catch, seed = 123)
}
\arguments{
\item{K}{<real> carrying capacity parameter.}

\item{r}{<real> intrinsic growth parameter.}

\item{q}{<real> relative index catchability.}

\item{P1}{<real> initial state relative to K (1 = B_t = K)}

\item{obs_std}{<real> lognormal standard deviation for relative index}

\item{pro_std}{<real> lognormal standard deviation for process equation}

\item{catch}{<vector> observed catches, defines number of years, so if you have a period of no catches set values of 0, number of years = length(catches+ 1)}

\item{seed}{<int> set.seed for reproducibility}
}
\description{
Returns a list of model quantities and observations that can be used as an operating model
when looking at different estimation models. A main idea here is changing the lognormal distribution to be more
traditional no adjustments.
}
