% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surplus_production_funs.R
\name{surplus_production_model_re_parameterised}
\alias{surplus_production_model_re_parameterised}
\title{surplus_production_model_re_parameterised}
\usage{
surplus_production_model_re_parameterised(log_jt, r, k, catches, catchability)
}
\arguments{
\item{log_jt}{<real> states specified as J_t = B_t / K}

\item{r}{<real> intrinsic growth parameter.}

\item{k}{<real> carrying capacity parameter.}

\item{catches}{<vector> observed catches, defines number of years, so if you have a period of no catches set values of 0, number of years = length(catches+ 1)}

\item{catchability}{<vector> a scalar to generate the relative index observation}

\item{n_t}{number of time steps}
}
\value{
a list of OM outputs
}
\description{
Returns a list of model quantities and observations that can be used as an operating model
this is more for helping me debug mu model
}
