% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/surplus_production_funs.R
\name{theta_logistic}
\alias{theta_logistic}
\title{theta_logistic}
\usage{
theta_logistic(N_t, sig_obs, sig_pro, r, K, theta, x0)
}
\arguments{
\item{N_t}{number of time steps}

\item{sig_obs}{= standard deviation to draw observations}

\item{sig_pro}{= standard deviation of process error}

\item{r}{<real> intrinsic growth parameter.}

\item{K}{<real> carrying capacity parameter.}

\item{theta}{}

\item{x0}{<real> initial state}
}
\value{
list of simulated data and derived quantities.
}
\description{
log-transformed theta logistic population growth model
(Wang, 2007)
}
