% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growth_params.R
\name{transpose_VB}
\alias{transpose_VB}
\title{transpose_VB}
\usage{
transpose_VB(L1, L2, Linf, k, t0)
}
\arguments{
\item{L1}{length at reference 1}

\item{L2}{length at reference 2, corresponds to g2}

\item{Linf}{mean asympote length}

\item{k}{VB growth rate}

\item{t0}{VB age at length 0}
}
\value{
vector<g1, g2> the linear growth increment model
}
\description{
transpose_VB
}
\details{
transpose the traditional VB formulation into a linear increment model based on the
paper \insertCite{francis1988maximum}{stockassessmenthelper}
}
\examples{
\dontrun{
transpose_VB(L1 = 50, L2 = 100, Linf = 155.9, k = 0.116, t0 = 2.495)
}
}
\references{
\insertAllCited{}
}
