% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_transformations.R
\name{ub_constrain}
\alias{ub_constrain}
\title{ub_constrain convert x(-Inf, Inf) -> y(-Inf, Ub]}
\usage{
ub_constrain(x, Ub)
}
\arguments{
\item{x}{value to transform}

\item{Ub}{upper bound of value}
}
\value{
constrained (natural space) value
}
\description{
ub_constrain convert x(-Inf, Inf) -> y(-Inf, Ub]
}
