% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_transformations.R
\name{ub_free}
\alias{ub_free}
\title{ub_free convert y(-Inf, Ub] -> x(-Inf, Inf)}
\usage{
ub_free(y, Ub)
}
\arguments{
\item{y}{value to transform}

\item{Ub}{upper bound of value}
}
\value{
unconstrained value
}
\description{
ub_free convert y(-Inf, Ub] -> x(-Inf, Inf)
}
