% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility_math_funs.R
\name{Var}
\alias{Var}
\title{overloading function to calculate variance when you have na's in vector}
\usage{
Var(..., na.rm = T)
}
\arguments{
\item{...}{same parameters as var(...)}

\item{na.rm}{a logical value indicating whether NA values should be stripped before the computation proceeds.}
}
\value{
variance with na's ignored
}
\description{
overloading function to calculate variance when you have na's in vector
}
