% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/param_transformations.R
\name{lb_free}
\alias{lb_free}
\title{lb_constrain convert y(Lb, Inf] -> x(-Inf, Inf)}
\usage{
lb_free(y, Lb)
}
\arguments{
\item{y}{value to transform}

\item{Lb}{Lower bound of value}
}
\value{
unconstrained value
}
\description{
lb_constrain convert y(Lb, Inf] -> x(-Inf, Inf)
}
