% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_comp_resids.R
\name{plot_comp_resids}
\alias{plot_comp_resids}
\title{plot_comp_resids}
\usage{
plot_comp_resids(resids_matrix, show_legend = TRUE)
}
\arguments{
\item{resids_matrix}{a matrix[years,bins] that are standardised residuals from compositional data, rownames(resids_matrix) = years (y-axis), colnames(resids_matrix) = ages/length}
}
\value{
a ggplot bubble plot
}
\description{
A function that generates a plot of residuals for compositional data.
}
\author{
C.Marsh
}
